Seperate HomeView to multiple subview
